Move site stuff to its own package
