+++ site/src/main/scala/org/typelevel/sbt/site/TypelevelProjects.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.typelevel.sbt
+package org.typelevel.sbt.site
 
 import sbt._
